fix(web-frontend): align link_row empty semantics with the backend
Replace isEmptyForRequiredValidation with a single isEmpty. A link_row is
a many-to-many relation, so the backend treats it as empty only when no row
is linked; the frontend's isEmpty was inheriting the base array check and
reporting a picked empty-primary row as empty. Make isEmpty selection-based
so the conditional-visibility filters and required form validation share one
definition that matches the backend.
